Introduce subcommands and cli argument parsing
In the future this tool is supposed to be able to generate typescript
types from an Onix xsd file. In order to distinguish between generating
types or generating the enums for the code lists, we are using
subcommands. Commander is a typescript package that allows us to
introduce subcommands and parsing of command line arguments. This change
introduces both, the parsing of command line arguments for the current
functionality (generating enums) as well as the introduction of
subcommands.