fix: generate enums correctly and add analysis
The enums have been generated with the incorrect enum names. The enum
names should be the refnames and not be generated from the annotation in
the codeList. This commit now generates the enums correctly, taking the
refname as the enum name.

The commit also adds a tool to analyse the bookProduct-xsd file by
generating a so called elementMap.json. The elementMap.json gives an
overview over the structure of all elements in the bookProduct.xsd.